Interface ServiceProviderInterface was moved from Silex to Pimple namespace since package silex 2.0.* (package pimple version is 3.0). Replace silex requirement by pimple (lot of silex dependencies are not needed). Change DerAlex\Silex namespace to DerAlex\Pimple for consistancy
