+++ src/DerAlex/Pimple/YamlConfigServiceProvider.php
@@ -20,10 +20,10 @@
  * FROM,  OUT OF  OR IN CONNECTION  WITH THE  SOFTWARE  OR THE  USE OR  OTHER *
  * DEALINGS IN THE SOFTWARE.                                                  *
  ******************************************************************************/
-namespace DerAlex\Silex;
+namespace DerAlex\Pimple;
 
-use Silex\Application;
-use Silex\ServiceProviderInterface;
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
 use Symfony\Component\Yaml\Yaml;
 
 
@@ -35,17 +35,20 @@ class YamlConfigServiceProvider implements ServiceProviderInterface
         $this->file = $file;
     }
 
-    public function register(Application $app) {
+    public function register(Container $pimple) {
         $config = Yaml::parse(file_get_contents($this->file));
 
         if (is_array($config)) {
-            $this->importSearch($config, $app);
+            $this->importSearch($config, $pimple);
 
-            if (isset($app['config']) && is_array($app['config'])) {
-                $app['config'] = array_replace_recursive($app['config'], $config);
+            if (isset($pimple['config']) && is_array($pimple['config'])) {
+                $pimple['config'] = array_replace_recursive($pimple['config'], $config);
             } else {
-                $app['config'] = $config;
+                $pimple['config'] = $config;
             }
+        } else {
+            // Consider YAML file malformed
+            $pimple['config'] = array();
         }
 
     }
@@ -56,20 +59,17 @@ class YamlConfigServiceProvider implements ServiceProviderInterface
      * @param array $config
      *   The result of Yaml::parse().
      */
-    public function importSearch(&$config, $app) {
+    public function importSearch(&$config, $pimple) {
         foreach ($config as $key => $value) {
             if ($key == 'imports') {
                 foreach ($value as $resource) {
                     $base_dir = str_replace(basename($this->file), '', $this->file);
                     $new_config = new YamlConfigServiceProvider($base_dir . $resource['resource']);
-                    $new_config->register($app);
+                    $new_config->register($pimple);
                 }
                 unset($config['imports']);
             }
         }
-    }
-
-    public function boot(Application $app) {
     }
 
     public function getConfigFile() {
